pyproject.toml
[project]
name = "certificate"
version = "0.1.0"
description = "The generic marimo notebook we use to generate and distribute our certificates."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "dotenv>=0.9.9",
  "marimo[recommended]>=0.17.7",
  "pillow>=12.0.0",
  "polars>=1.35.1",
  "yagmail>=0.15.293",
]

[dependency-groups]
dev = ["ruff>=0.14.4"]
